Conflicts in src/modbus-rtu.c resolved in favour of upstream:

* stop bits are set in _modbus_rtu_connect() along with the other port
  parameters again, so the extra _modbus_rtu_set_stopbits_onthefly() call
  there is dropped
* the debug printf in the WIN32 branch no longer prints stop_bit_receive,
  different stop bits are POSIX-only

+++ tests/bandwidth-client.c
@@ -1,18 +1,19 @@
 /*
- * Copyright © 2008-2014 Stéphane Raimbault <stephane.raimbault@gmail.com>
+ * Copyright © Stéphane Raimbault <stephane.raimbault@gmail.com>
  *
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
 #include <stdio.h>
 #ifndef _MSC_VER
-#include <unistd.h>
 #include <sys/time.h>
+#include <unistd.h>
 #endif
-#include <string.h>
-#include <stdlib.h>
-#include <time.h>
 #include <errno.h>
+#include <inttypes.h>
+#include <stdlib.h>
+#include <string.h>
+#include <time.h>
 
 #include <modbus.h>
 
@@ -46,7 +47,7 @@ int main(int argc, char *argv[])
     uint32_t start;
     uint32_t end;
     uint32_t bytes;
-    uint32_t rate;
+    uint64_t rate;
     int rc;
     int n_loop;
     int use_backend;
@@ -59,7 +60,8 @@ int main(int argc, char *argv[])
             use_backend = RTU;
             n_loop = 100;
         } else {
-            printf("Usage:\n  %s [tcp|rtu] - Modbus client to measure data bandwidth\n\n", argv[0]);
+            printf("Usage:\n  %s [tcp|rtu] - Modbus client to measure data bandwidth\n\n",
+                   argv[0]);
             exit(1);
         }
     } else {
@@ -75,8 +77,7 @@ int main(int argc, char *argv[])
         modbus_set_slave(ctx, 1);
     }
     if (modbus_connect(ctx) == -1) {
-        fprintf(stderr, "Connection failed: %s\n",
-                modbus_strerror(errno));
+        fprintf(stderr, "Connection failed: %s\n", modbus_strerror(errno));
         modbus_free(ctx);
         return -1;
     }
@@ -93,7 +94,7 @@ int main(int argc, char *argv[])
 
     nb_points = MODBUS_MAX_READ_BITS;
     start = gettime_ms();
-    for (i=0; i<n_loop; i++) {
+    for (i = 0; i < n_loop; i++) {
         rc = modbus_read_bits(ctx, 0, nb_points, tab_bit);
         if (rc == -1) {
             fprintf(stderr, "%s\n", modbus_strerror(errno));
@@ -103,17 +104,17 @@ int main(int argc, char *argv[])
     end = gettime_ms();
     elapsed = end - start;
 
-    rate = (n_loop * nb_points) * G_MSEC_PER_SEC / (end - start);
+    rate = (uint64_t) n_loop * nb_points * G_MSEC_PER_SEC / (end - start);
     printf("Transfer rate in points/seconds:\n");
-    printf("* %d points/s\n", rate);
+    printf("* %" PRIu64 " points/s\n", rate);
     printf("\n");
 
-    bytes = n_loop * (nb_points / 8) + ((nb_points % 8) ? 1 : 0);
-    rate = bytes / 1024 * G_MSEC_PER_SEC / (end - start);
+    bytes = n_loop * ((nb_points / 8) + ((nb_points % 8) ? 1 : 0));
+    rate = (uint64_t) bytes / 1024 * G_MSEC_PER_SEC / (end - start);
     printf("Values:\n");
     printf("* %d x %d values\n", n_loop, nb_points);
     printf("* %.3f ms for %d bytes\n", elapsed, bytes);
-    printf("* %d KiB/s\n", rate);
+    printf("* %" PRIu64 " KiB/s\n", rate);
     printf("\n");
 
     /* TCP: Query and response header and values */
@@ -121,16 +122,16 @@ int main(int argc, char *argv[])
     printf("Values and TCP Modbus overhead:\n");
     printf("* %d x %d bytes\n", n_loop, bytes);
     bytes = n_loop * bytes;
-    rate = bytes / 1024 * G_MSEC_PER_SEC / (end - start);
+    rate = (uint64_t) bytes / 1024 * G_MSEC_PER_SEC / (end - start);
     printf("* %.3f ms for %d bytes\n", elapsed, bytes);
-    printf("* %d KiB/s\n", rate);
+    printf("* %" PRIu64 " KiB/s\n", rate);
     printf("\n\n");
 
     printf("READ REGISTERS\n\n");
 
     nb_points = MODBUS_MAX_READ_REGISTERS;
     start = gettime_ms();
-    for (i=0; i<n_loop; i++) {
+    for (i = 0; i < n_loop; i++) {
         rc = modbus_read_registers(ctx, 0, nb_points, tab_reg);
         if (rc == -1) {
             fprintf(stderr, "%s\n", modbus_strerror(errno));
@@ -140,17 +141,17 @@ int main(int argc, char *argv[])
     end = gettime_ms();
     elapsed = end - start;
 
-    rate = (n_loop * nb_points) * G_MSEC_PER_SEC / (end - start);
+    rate = (uint64_t) n_loop * nb_points * G_MSEC_PER_SEC / (end - start);
     printf("Transfer rate in points/seconds:\n");
-    printf("* %d registers/s\n", rate);
+    printf("* %" PRIu64 " registers/s\n", rate);
     printf("\n");
 
     bytes = n_loop * nb_points * sizeof(uint16_t);
-    rate = bytes / 1024 * G_MSEC_PER_SEC / (end - start);
+    rate = (uint64_t) bytes / 1024 * G_MSEC_PER_SEC / (end - start);
     printf("Values:\n");
     printf("* %d x %d values\n", n_loop, nb_points);
     printf("* %.3f ms for %d bytes\n", elapsed, bytes);
-    printf("* %d KiB/s\n", rate);
+    printf("* %" PRIu64 " KiB/s\n", rate);
     printf("\n");
 
     /* TCP:Query and response header and values */
@@ -158,19 +159,18 @@ int main(int argc, char *argv[])
     printf("Values and TCP Modbus overhead:\n");
     printf("* %d x %d bytes\n", n_loop, bytes);
     bytes = n_loop * bytes;
-    rate = bytes / 1024 * G_MSEC_PER_SEC / (end - start);
+    rate = (uint64_t) bytes / 1024 * G_MSEC_PER_SEC / (end - start);
     printf("* %.3f ms for %d bytes\n", elapsed, bytes);
-    printf("* %d KiB/s\n", rate);
+    printf("* %" PRIu64 " KiB/s\n", rate);
     printf("\n\n");
 
     printf("WRITE AND READ REGISTERS\n\n");
 
     nb_points = MODBUS_MAX_WR_WRITE_REGISTERS;
     start = gettime_ms();
-    for (i=0; i<n_loop; i++) {
-        rc = modbus_write_and_read_registers(ctx,
-                                             0, nb_points, tab_reg,
-                                             0, nb_points, tab_reg);
+    for (i = 0; i < n_loop; i++) {
+        rc = modbus_write_and_read_registers(
+            ctx, 0, nb_points, tab_reg, 0, nb_points, tab_reg);
         if (rc == -1) {
             fprintf(stderr, "%s\n", modbus_strerror(errno));
             return -1;
@@ -179,17 +179,17 @@ int main(int argc, char *argv[])
     end = gettime_ms();
     elapsed = end - start;
 
-    rate = (n_loop * nb_points) * G_MSEC_PER_SEC / (end - start);
+    rate = (uint64_t) n_loop * nb_points * G_MSEC_PER_SEC / (end - start);
     printf("Transfer rate in points/seconds:\n");
-    printf("* %d registers/s\n", rate);
+    printf("* %" PRIu64 " registers/s\n", rate);
     printf("\n");
 
     bytes = n_loop * nb_points * sizeof(uint16_t);
-    rate = bytes / 1024 * G_MSEC_PER_SEC / (end - start);
+    rate = (uint64_t) bytes / 1024 * G_MSEC_PER_SEC / (end - start);
     printf("Values:\n");
     printf("* %d x %d values\n", n_loop, nb_points);
     printf("* %.3f ms for %d bytes\n", elapsed, bytes);
-    printf("* %d KiB/s\n", rate);
+    printf("* %" PRIu64 " KiB/s\n", rate);
     printf("\n");
 
     /* TCP:Query and response header and values */
@@ -197,9 +197,9 @@ int main(int argc, char *argv[])
     printf("Values and TCP Modbus overhead:\n");
     printf("* %d x %d bytes\n", n_loop, bytes);
     bytes = n_loop * bytes;
-    rate = bytes / 1024 * G_MSEC_PER_SEC / (end - start);
+    rate = (uint64_t) bytes / 1024 * G_MSEC_PER_SEC / (end - start);
     printf("* %.3f ms for %d bytes\n", elapsed, bytes);
-    printf("* %d KiB/s\n", rate);
+    printf("* %" PRIu64 " KiB/s\n", rate);
     printf("\n");
 
     /* Free the memory */
